fixed warnings in linux and no mem leaks

+++ week12/backend/print_vector.hpp
@@ -8,7 +8,7 @@ using namespace std;
 
 template<typename V>
 void print_vec(vector<V>& v){
-    for (int i = 0 ; i < v.size() ; i ++ ){
+    for (int i = 0 ; i < (int)v.size() ; i ++ ){
         cout << v[i] << " ";
     }
     cout << endl;
